// src/types/notify.ts
export interface SiteMessageDto {
  readonly id: string;
  readonly recordId: string;
  readonly receiverUserId: string;
  readonly title: string;
  readonly content: string;
  readonly readStatus: boolean;
  readonly readTime?: string;
  readonly createTime?: string;
  readonly updateTime?: string;
}
// ...
export type SiteMessageResp = SiteMessageDto;